Added main menu with options
