+++ Assets/Art/Font/UIFont.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ceb067532853fd04790cf60849632c8b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
