+++ Assets/Dev/Scripts/Managers/Gamemanager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 667f03280a1e2674d8d4d80587089052
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
