+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,4 +11,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/MainMenu.unity
     guid: bd46e6de7fa5fb44880bd17adea16168
+  - enabled: 1
+    path: Assets/Scenes/Straatroof Scenario.unity
+    guid: 22fade062edaa024a997d3794c618812
   m_configObjects: {}
